fix: fix instance cache managing
