+++ src/5_entities/instances/api/index.ts
@@ -1,6 +1,2 @@
 export * from './rawApi';
-export * from './instance';
-export * from './process';
-export * from './utils';
-export * from './config';
-export * from './content';
+export * from './tanstack';
