+++ src/5_entities/instances/api/tanstack/config.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/solid-query';
+import { getImportConfigsRaw } from '../rawApi';
+import { CONFIG_QUERY_KEYS } from './config_query_keys';
+
+export const useImportConfigs = () => {
+  return useQuery(() => ({
+    queryKey: CONFIG_QUERY_KEYS.IMPORT(),
+    queryFn: getImportConfigsRaw,
+  }));
+};
