+++ src/5_entities/instances/api/tanstack/config_query_keys.ts
@@ -1,0 +1,4 @@
+export const CONFIG_QUERY_KEYS = {
+  SELF: ['config'],
+  IMPORT: () => [...CONFIG_QUERY_KEYS.SELF, 'import'],
+} as const;
