+++ src/5_entities/instances/api/tanstack/instance_query_keys.ts
@@ -1,0 +1,10 @@
+export const INSTANCE_QUERY_KEYS = {
+  SELF: ['instance'],
+  LIST: () => [...INSTANCE_QUERY_KEYS.SELF, 'list'],
+  GET: (id: string) => [...INSTANCE_QUERY_KEYS.SELF, 'get', id],
+  DIR: (id: string) => [...INSTANCE_QUERY_KEYS.GET(id), 'dir'],
+  PROCESS: {
+    SELF: () => [...INSTANCE_QUERY_KEYS.SELF, 'process'],
+    BY_INSTANCE: (id: string) => [...INSTANCE_QUERY_KEYS.PROCESS.SELF(), id],
+  },
+} as const;
