+++ src/5_entities/instances/api/tanstack/utils.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/solid-query';
-import { revealInExplorerRaw } from './rawApi';
+import { revealInExplorerRaw } from '../rawApi';
 import { useTranslation } from '@/shared/model';
 import { showError } from '@/shared/lib/showError';
